+++ Slam/src/main/java/com/system/slam/entity/Category.java
@@ -14,7 +14,7 @@ public class Category {
     @Id
     @Column(name = "IdCategory", nullable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idCategory;
+    private Long idCategory;
 
     @Column(name = "Name", length = 25, nullable = false)
     private String name;
